refactor: move helpers to utils.py
